--- Assets/Scripts/ExtentionMethods/GameObjectExtensions.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public static class GameObjectExtensions
-{
-    public static bool IsInLayerMask(this GameObject obj, LayerMask mask)
-    {
-        return ((mask.value & (1 << obj.layer)) > 0);
-    }
-}
