+++ Assets/Scripts/UI/ScaleWithInspector.cs
@@ -1,4 +1,4 @@
-using SLIDDES.UI;
+using Netherlands3D.Twin.ExtensionMethods;
 using UnityEngine;
 
 namespace Netherlands3D.Twin
