+++ Assets/_Application/Cameras/Input/CameraInputSystemProvider.cs
@@ -1,6 +1,7 @@
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
+using Netherlands3D.Twin.ExtensionMethods;
 using UnityEngine;
 using UnityEngine.EventSystems;
 #if ENABLE_INPUT_SYSTEM
