+++ Assets/_Application/Layers/Properties/Properties.cs
@@ -1,3 +1,4 @@
+using Netherlands3D.Twin.ExtensionMethods;
 using Netherlands3D.Twin.Layers.LayerTypes;
 using UnityEngine;
