+++ Assets/_Functionalities/OGC3DTiles/Scripts/Tiles3DContentOverlayInspector.cs
@@ -1,5 +1,5 @@
-using Netherlands3D.Twin;
 using Netherlands3D.Credentials;
+using Netherlands3D.Twin.ExtensionMethods;
 using Netherlands3D.Twin.Layers;
 using Netherlands3D.Twin.Layers.LayerTypes.Credentials.Properties;
 using Netherlands3D.Twin.Layers.UI.HierarchyInspector;
